+++ examples/plot_discriminant_scores.py
@@ -22,17 +22,46 @@ is_light = df["HadronConeExclTruthLabelID"] == 0
 is_c = df["HadronConeExclTruthLabelID"] == 4
 is_b = df["HadronConeExclTruthLabelID"] == 5
 
-hist_dips_light = Histogram(df[is_light]["disc_dips"], flavour="ujets", label="DIPS")
-hist_dips_c = Histogram(df[is_c]["disc_dips"], flavour="cjets", label="DIPS")
-hist_dips_b = Histogram(df[is_b]["disc_dips"], flavour="bjets", label="DIPS")
+hist_dips_light = Histogram(
+    df[is_light]["disc_dips"],
+    flavour="ujets",
+    label="DIPS",
+    # if not set, the "ratio_group" argument would be set to the flavour "ujets"
+    # so this is just done here for demonstration purposes
+    ratio_group="ujets",
+)
+hist_dips_c = Histogram(
+    df[is_c]["disc_dips"],
+    flavour="cjets",
+    label="DIPS",
+    ratio_group="cjets",
+)
+hist_dips_b = Histogram(
+    df[is_b]["disc_dips"],
+    flavour="bjets",
+    label="DIPS",
+    ratio_group="bjets",
+)
 hist_rnnip_light = Histogram(
-    df[is_light]["disc_rnnip"], flavour="ujets", label="RNNIP", linestyle="dashed"
+    df[is_light]["disc_rnnip"],
+    flavour="ujets",
+    label="RNNIP",
+    linestyle="dashed",
+    ratio_group="ujets",
 )
 hist_rnnip_c = Histogram(
-    df[is_c]["disc_rnnip"], flavour="cjets", label="RNNIP", linestyle="dashed"
+    df[is_c]["disc_rnnip"],
+    flavour="cjets",
+    label="RNNIP",
+    linestyle="dashed",
+    ratio_group="cjets",
 )
 hist_rnnip_b = Histogram(
-    df[is_b]["disc_rnnip"], flavour="bjets", label="RNNIP", linestyle="dashed"
+    df[is_b]["disc_rnnip"],
+    flavour="bjets",
+    label="RNNIP",
+    linestyle="dashed",
+    ratio_group="bjets",
 )
 
 # Initialise histogram plot